created circuit_library_utils.py + tests
